refactor(user): reorganize user's routers
