+++ src/routers/index.js
@@ -1,14 +1,4 @@
 /**
- * 前往用户页面
- * @param id 用户id
- */
-export function toUserPage(id) {
-  uni.navigateTo({
-    url: `/pages/users/detail?id=${id}`,
-  });
-}
-
-/**
  * 跳转至首页
  */
 export function toIndexPage(closeAll = false) {
@@ -24,57 +14,11 @@ export function toIndexPage(closeAll = false) {
 }
 
 /**
- * 前往用户信息页面
- */
-export function toUserInfoPage() {
-  uni.navigateTo({
-    url: '/pages/about/userinfo/userinfo',
-  });
-}
-
-/**
- * 前往用户贡献语音页面
- * @param id 用户id
- */
-export function toMyRecordsPage(id) {
-  uni.navigateTo({
-    url: `/pages/users/pronunciation?id=${id}`,
-  });
-}
-
-/**
  * 前往贡献语音榜单页面
  */
 export function toVoiceRankListPage() {
   uni.navigateTo({
     url: '/pages/component/VoiceRankList/VoiceRankList',
-  });
-}
-
-/**
- * 前往更改用户昵称页面
- */
-export function toChangeNicknamePage() {
-  uni.navigateTo({
-    url: '/pages/about/nickname/nickname',
-  });
-}
-
-/**
- * 前往更改邮箱页面
- */
-export function toChangeEmailPage() {
-  uni.navigateTo({
-    url: '/pages/about/email/email',
-  });
-}
-
-/**
- * 前往更改手机号页面
- */
-export function toChangePhonePage() {
-  uni.navigateTo({
-    url: '/pages/about/phone/phone',
   });
 }
 
@@ -139,18 +83,6 @@ export function toSearchQuizPage() {
   uni.navigateTo({
     url: '/pages/tools/quiz/SearchQuiz/SearchQuiz',
   });
-}
-
-export function toMePage(closeAll = false) {
-  if (closeAll) {
-    uni.reLaunch({
-      url: '/pages/index/index?status=me',
-    });
-  } else {
-    uni.navigateTo({
-      url: '/pages/index/index?status=me',
-    });
-  }
 }
 
 /**
